Added transparency suport for Web view.

+++ sources/plug_Awesomium/include/luna/wrappers/wrapper_Awesomium_WebCore.h
@@ -52,7 +52,7 @@ public:
 		_obj.pushArg(width);
 		_obj.pushArg(height);
 		_obj.pushArg(session);
-		_obj.pushArg(type);
+		_obj.pushArg((int)type);
 		return (_obj.callFunction<Awesomium::WebView*>());
 	};
